+++ lib/lightning/extensions/collection_hook.ex
@@ -1,0 +1,18 @@
+defmodule Lightning.Extensions.CollectionHook do
+  @moduledoc """
+  Callbacks for additional processing on collections operations.
+  """
+  @behaviour Lightning.Extensions.CollectionHooking
+
+  @impl true
+  def handle_create(_attrs), do: :ok
+
+  @impl true
+  def handle_delete(_col), do: :ok
+
+  @impl true
+  def handle_put_items(_col, _size), do: :ok
+
+  @impl true
+  def handle_delete_items(_col, _size), do: :ok
+end